Evito autoboxing en els tests per millorar llegibilitat
git-svn-id: http://devel.cpl.upc.edu/recursos/svn/smeagol/java/branches/t251@1406 4ab63a03-c2d3-4037-8559-9a21b594da05

+++ smeagol-client/src/test/java/edu/upc/cpl/smeagol/client/domain/RecurrentBookingTest.java
@@ -18,19 +18,19 @@ import edu.upc.cpl.smeagol.client.ical.Frequency;
 public class RecurrentBookingTest extends TestCase {
 	private static Logger logger = Logger.getLogger(RecurrentBookingTest.class);
 
-	private static long ID_RESOURCE1 = 111;
-	private static long ID_EVENT1 = 222;
+	private static Long ID_RESOURCE1 = 111L;
+	private static Long ID_EVENT1 = 222L;
 	private static DateTime DTSTART1 = new DateTime(2011, 2, 3, 8, 0, 0, 0);
 	private static DateTime DTEND1 = new DateTime(2011, 2, 3, 10, 30, 0, 0);
-	private static short INTERVAL1 = 2;
+	private static Short INTERVAL1 = 2;
 	private static DateTime UNTIL1 = new DateTime(2011, 7, 9, 0, 0, 0, 0);
 	private static Frequency FREQUENCY1 = Frequency.WEEKLY;
 	private static Set<DayOfWeek> BY_DAY1 = new HashSet<DayOfWeek>();
 	private static Set<Short> BY_DAYOFMONTH1 = new TreeSet<Short>();
 	private static Set<Short> BY_MONTH1 = new TreeSet<Short>();
 
-	private static long ID_RESOURCE2 = 2;
-	private static long ID_EVENT2 = 20;
+	private static Long ID_RESOURCE2 = 2L;
+	private static Long ID_EVENT2 = 20L;
 	private static DateTime DTSTART2 = new DateTime("2011-03-02T15:00:00");
 	private static DateTime DTEND2 = new DateTime("2011-03-02T16:00:00");
 	private static RecurrentBooking B1;
@@ -69,11 +69,11 @@ public class RecurrentBookingTest extends TestCase {
 		RecurrentBooking r = RecurrentBooking.asDailyRecurrence(ID_RESOURCE1, ID_EVENT1, DTSTART1, DTEND1, INTERVAL1,
 				UNTIL1);
 		assertEquals(Frequency.DAILY, r.getFrequency());
-		assertEquals(ID_RESOURCE1, r.getIdResource().longValue());
-		assertEquals(ID_EVENT1, r.getIdEvent().longValue());
+		assertEquals(ID_RESOURCE1, r.getIdResource());
+		assertEquals(ID_EVENT1, r.getIdEvent());
 		assertNotNull(r.getDtStart());
 		assertEquals(DTEND1, r.getDtEnd());
-		assertEquals(INTERVAL1, r.getInterval().shortValue());
+		assertEquals(INTERVAL1, r.getInterval());
 		assertEquals(UNTIL1, r.getUntil());
 	}
 
@@ -81,8 +81,8 @@ public class RecurrentBookingTest extends TestCase {
 	public void testAsDailyRecurrenceWithDefaultArgs() {
 		RecurrentBooking r = RecurrentBooking.asDailyRecurrence(ID_RESOURCE1, ID_EVENT1, null, null, null, null);
 		assertEquals(Frequency.DAILY, r.getFrequency());
-		assertEquals(ID_RESOURCE1, r.getIdResource().longValue());
-		assertEquals(ID_EVENT1, r.getIdEvent().longValue());
+		assertEquals(ID_RESOURCE1, r.getIdResource());
+		assertEquals(ID_EVENT1, r.getIdEvent());
 		assertNotNull(r.getDtStart());
 		assertEquals(r.getDtStart().plusHours(1), r.getDtEnd());
 		assertEquals(1, r.getInterval().shortValue());
@@ -94,11 +94,11 @@ public class RecurrentBookingTest extends TestCase {
 		RecurrentBooking r = RecurrentBooking.asWeeklyRecurrence(ID_RESOURCE1, ID_EVENT1, DTSTART1, DTEND1, INTERVAL1,
 				UNTIL1, BY_DAY1);
 		assertEquals(Frequency.WEEKLY, r.getFrequency());
-		assertEquals(ID_RESOURCE1, r.getIdResource().longValue());
-		assertEquals(ID_EVENT1, r.getIdEvent().longValue());
+		assertEquals(ID_RESOURCE1, r.getIdResource());
+		assertEquals(ID_EVENT1, r.getIdEvent());
 		assertNotNull(r.getDtStart());
 		assertEquals(DTEND1, r.getDtEnd());
-		assertEquals(INTERVAL1, r.getInterval().shortValue());
+		assertEquals(INTERVAL1, r.getInterval());
 		assertEquals(UNTIL1, r.getUntil());
 		assertEquals(BY_DAY1, r.getByDay());
 	}
@@ -107,8 +107,8 @@ public class RecurrentBookingTest extends TestCase {
 	public void testAsWeeklyRecurrenceWithDefaultArgs() {
 		RecurrentBooking r = RecurrentBooking.asWeeklyRecurrence(ID_RESOURCE1, ID_EVENT1, null, null, null, null, null);
 		assertEquals(Frequency.WEEKLY, r.getFrequency());
-		assertEquals(ID_RESOURCE1, r.getIdResource().longValue());
-		assertEquals(ID_EVENT1, r.getIdEvent().longValue());
+		assertEquals(ID_RESOURCE1, r.getIdResource());
+		assertEquals(ID_EVENT1, r.getIdEvent());
 		assertNotNull(r.getDtStart());
 		assertEquals(r.getDtStart().plusHours(1), r.getDtEnd());
 		assertEquals(1, r.getInterval().shortValue());
@@ -122,11 +122,11 @@ public class RecurrentBookingTest extends TestCase {
 		RecurrentBooking r = RecurrentBooking.asMonthlyRecurrence(ID_RESOURCE1, ID_EVENT1, DTSTART1, DTEND1, INTERVAL1,
 				UNTIL1, BY_DAYOFMONTH1);
 		assertEquals(Frequency.MONTHLY, r.getFrequency());
-		assertEquals(ID_RESOURCE1, r.getIdResource().longValue());
-		assertEquals(ID_EVENT1, r.getIdEvent().longValue());
+		assertEquals(ID_RESOURCE1, r.getIdResource());
+		assertEquals(ID_EVENT1, r.getIdEvent());
 		assertNotNull(r.getDtStart());
 		assertEquals(DTEND1, r.getDtEnd());
-		assertEquals(INTERVAL1, r.getInterval().shortValue());
+		assertEquals(INTERVAL1, r.getInterval());
 		assertEquals(UNTIL1, r.getUntil());
 		assertEquals(BY_DAYOFMONTH1, r.getByDayOfMonth());
 	}
@@ -136,8 +136,8 @@ public class RecurrentBookingTest extends TestCase {
 		RecurrentBooking r = RecurrentBooking
 				.asMonthlyRecurrence(ID_RESOURCE1, ID_EVENT1, null, null, null, null, null);
 		assertEquals(Frequency.MONTHLY, r.getFrequency());
-		assertEquals(ID_RESOURCE1, r.getIdResource().longValue());
-		assertEquals(ID_EVENT1, r.getIdEvent().longValue());
+		assertEquals(ID_RESOURCE1, r.getIdResource());
+		assertEquals(ID_EVENT1, r.getIdEvent());
 		assertNotNull(r.getDtStart());
 		assertEquals(r.getDtStart().plusHours(1), r.getDtEnd());
 		assertEquals(1, r.getInterval().shortValue());
@@ -151,11 +151,11 @@ public class RecurrentBookingTest extends TestCase {
 		RecurrentBooking r = RecurrentBooking.asYearlyRecurrence(ID_RESOURCE1, ID_EVENT1, DTSTART1, DTEND1, INTERVAL1,
 				UNTIL1, BY_DAYOFMONTH1, BY_MONTH1);
 		assertEquals(Frequency.YEARLY, r.getFrequency());
-		assertEquals(ID_RESOURCE1, r.getIdResource().longValue());
-		assertEquals(ID_EVENT1, r.getIdEvent().longValue());
+		assertEquals(ID_RESOURCE1, r.getIdResource());
+		assertEquals(ID_EVENT1, r.getIdEvent());
 		assertNotNull(r.getDtStart());
 		assertEquals(DTEND1, r.getDtEnd());
-		assertEquals(INTERVAL1, r.getInterval().shortValue());
+		assertEquals(INTERVAL1, r.getInterval());
 		assertEquals(UNTIL1, r.getUntil());
 		assertEquals(BY_DAYOFMONTH1, r.getByDayOfMonth());
 	}
@@ -165,8 +165,8 @@ public class RecurrentBookingTest extends TestCase {
 		RecurrentBooking r = RecurrentBooking.asYearlyRecurrence(ID_RESOURCE1, ID_EVENT1, null, null, null, null, null,
 				null);
 		assertEquals(Frequency.YEARLY, r.getFrequency());
-		assertEquals(ID_RESOURCE1, r.getIdResource().longValue());
-		assertEquals(ID_EVENT1, r.getIdEvent().longValue());
+		assertEquals(ID_RESOURCE1, r.getIdResource());
+		assertEquals(ID_EVENT1, r.getIdEvent());
 		assertNotNull(r.getDtStart());
 		assertEquals(r.getDtStart().plusHours(1), r.getDtEnd());
 		assertEquals(1, r.getInterval().shortValue());
